requirement.txt
# Core Kafka packages
kafka-python==2.0.2
confluent-kafka==2.3.0

# Schema Registry support
confluent-kafka[avro]==2.3.0
confluent-kafka[schema-registry]==2.3.0
avro-python3==1.10.2
fastavro==1.9.4

# Alternative Kafka clients (choose one)
# aiokafka==0.10.0  # For async operations
# pykafka==2.8.0    # Alternative client

# Data processing
pandas==2.1.4
numpy==1.24.4

# YAML support
PyYAML==6.0.1
ruamel.yaml==0.18.5  # More advanced YAML parsing with comments preservation

# JSON Schema validation (often used with Schema Registry)
jsonschema==4.20.0

# Configuration management
python-dotenv==1.0.0
configparser==6.0.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# HTTP client for Schema Registry API calls
requests==2.31.0
urllib3==2.1.0

# Serialization formats
protobuf==4.25.1
msgpack==1.0.7

# Development and testing
pytest==7.4.4
pytest-asyncio==0.23.2
mock==5.1.0

# Optional: Database connectors (if storing processed data)
sqlalchemy
mysql-connector-python
# psycopg2-binary==2.9.9  # PostgreSQL
pymongo        # MongoDB

# Optional: Cloud integrations
# boto3==1.34.34          # AWS
# azure-storage-blob==12.19.0  # Azure
# google-cloud-storage==2.10.0  # GCP

# Optional: Additional data formats
# pyarrow==14.0.2         # Parquet support
# openpyxl==3.1.2         # Excel files
# xlrd==2.0.1             # Excel files (legacy)

# Security and authentication
# cryptography==41.0.8
# pyjwt==2.8.0

# Utilities
click==8.1.7              # CLI interface
tqdm==4.66.1              # Progress bars
python-dateutil==2.8.2   # Date parsing
pytz==2023.4              # Timezone support

#google cloud 
#google-cloud-sql-python-connector 
#google-auth


#For Synthetic data
Faker